Implemented data deletion for table schema deletion. Refs #292607

+++ dataset-service/src/main/java/org/eea/dataset/controller/DatasetSchemaControllerImpl.java
@@ -4,6 +4,7 @@ import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import com.netflix.hystrix.contrib.javanica.annotation.HystrixProperty;
 import io.netty.util.internal.StringUtil;
 import io.swagger.annotations.*;
+import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.eea.dataset.service.*;
 import org.eea.exception.EEAErrorMessage;
@@ -19,6 +20,7 @@ import org.eea.interfaces.vo.communication.UserNotificationContentVO;
 import org.eea.interfaces.vo.communication.UserNotificationVO;
 import org.eea.interfaces.vo.dataflow.DataFlowVO;
 import org.eea.interfaces.vo.dataflow.enums.TypeStatusEnum;
+import org.eea.interfaces.vo.dataset.DataSetMetabaseVO;
 import org.eea.interfaces.vo.dataset.DesignDatasetVO;
 import org.eea.interfaces.vo.dataset.OrderVO;
 import org.eea.interfaces.vo.dataset.enums.DataType;
@@ -37,6 +39,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -119,9 +122,9 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
   @Autowired
   private NotificationControllerZuul notificationControllerZuul;
 
-  /** The kafka sender utils. */
+  @Lazy
   @Autowired
-  private KafkaSenderUtils kafkaSenderUtils;
+  private BigDataDatasetService bigDataDatasetService;
 
   /**
    * Creates the empty dataset schema.
@@ -613,7 +616,7 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
           @ApiParam(type = "Long", value = "Dataset Id",
                   example = "0") @PathVariable("datasetId") Long datasetId,
           @ApiParam(type = "String", value = "table Schema Id",
-                  example = "5cf0e9b3b793310e9ceca190") @PathVariable("tableSchemaId") String tableSchemaId) {
+                  example = "5cf0e9b3b793310e9ceca190") @PathVariable("tableSchemaId") String tableSchemaId) throws Exception {
 
     if (!TypeStatusEnum.DESIGN.equals(dataflowControllerZuul
             .getMetabaseById(datasetService.getDataFlowIdById(datasetId)).getStatus())) {
@@ -622,8 +625,16 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
 
     try {
       final String datasetSchemaId = dataschemaService.getDatasetSchemaId(datasetId);
+      DataSetMetabaseVO dataSetMetabaseVO = datasetMetabaseService.findDatasetMetabase(datasetId);
+      Boolean isBigData = dataflowControllerZuul.isBigDataflow(dataSetMetabaseVO.getDataflowId());
 
       LOG.info("Deleting table schema with id {} for datasetId {}",tableSchemaId, datasetId);
+
+      //if table is big data remove first data from s3
+      if(BooleanUtils.isTrue(isBigData)){
+        bigDataDatasetService.deleteTableData(datasetId, dataSetMetabaseVO.getDataflowId(), dataSetMetabaseVO.getDataProviderId(), tableSchemaId, null, false);
+      }
+
       // Delete the Pk if needed from the catalogue, for all the fields of the table
       dataschemaService.deleteFromPkCatalogue(datasetSchemaId, tableSchemaId, datasetId);
 
@@ -635,9 +646,12 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
       // we delete the rules associate to the table
       rulesControllerZuul.deleteRuleByReferenceId(datasetSchemaId, tableSchemaId);
 
-      datasetService.deleteTableValue(datasetId, tableSchemaId);
+      //if table is not big data remove citus values and update materialized views
+      if(!BooleanUtils.isTrue(isBigData)){
+        datasetService.deleteTableValue(datasetId, tableSchemaId);
+        recordStoreControllerZuul.createUpdateQueryView(datasetId, false);
+      }
 
-      recordStoreControllerZuul.createUpdateQueryView(datasetId, false);
       LOG.info("Successfully deleted table schema with id {} for datasetId {}",tableSchemaId, datasetId);
     } catch (EEAException e) {
       LOG.error("Error deleting table schema with id {} for datasetId {} Message: {}", tableSchemaId, datasetId, e.getMessage(), e);
